requirements.txt
asn1crypto==1.5.1
boto3==1.40.26
botocore==1.40.26
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dotenv==0.9.9
filelock==3.19.1
fonttools==4.60.0
idna==3.10
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.13.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
snowflake-connector-python==3.17.3
snowflake-snowpark-python==1.39.0
sortedcontainers==2.4.0
threadpoolctl==3.6.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2,<3
